+++ fiddles/angular2/fiddle-0020-LayoutContainer/src/center/ScrollBox.js
@@ -1,1 +1,0 @@
-//# sourceMappingURL=ScrollBox.js.map