+++ fiddles/angular2/fiddle-0020-LayoutContainer/src/navigation/Navigation.js
@@ -1,1 +1,0 @@
-//# sourceMappingURL=Navigation.js.map