+++ fiddles/angular2/fiddle-0022-MicroApp/src/MicroApp.ts
@@ -1,0 +1,17 @@
+import {Component} from 'angular2/core';
+import * as core from 'angular2/core';
+import {Fiddle} from './component/Fiddle';
+
+@Component({
+    selector: 'app',
+    template: `
+   <fiddle>
+      <div head>HEADER</div>
+      <div menu>MENU</div>
+      <div body>BODY</div>
+   </fiddle>
+  `,
+    directives: [Fiddle]
+})
+export class MicroApp {
+}