+++ fiddles/angular2/fiddle-0022-MicroApp/src/component/Fiddle.js
@@ -1,0 +1,1 @@
+//# sourceMappingURL=Fiddle.js.map