+++ fiddles/angular2/fiddle-0024-DashboardLayout/src/meta.js
@@ -1,0 +1,9 @@
+var metadata = {
+    fiddleHeader: 'Angular 2 - Dashboard',
+    urls: {
+        data: 'https://s3-us-west-2.amazonaws.com/s.cdpn.io/297733/financial-800.json',
+        github: 'https://github.com/bradyhouse/house/tree/master/fiddles/angular2/fiddle-0024-Dashboard'
+    },
+    consoleTag: 'H O U S E ~ f i d d l e s'
+};
+//# sourceMappingURL=meta.js.map